requirements.txt
Flask
psycopg2-binary
torch
transformers
sentence-transformers
pandas
numpy
scikit-learn
torchvision
huggingface_hub
Pillow
firebase-admin